# swagger/LibraryTestTask.json
{"openapi":"3.0.1","info":{"title":"Library Api","description":"Test task","contact":{"name":"Azarkin Vitaliy","url":"https://github.com/Azark1n","email":"v.azarkin@gmail.com"},"version":"1.0"},"servers":[{"url":"http://nr.ddns.ms:8013","description":"Generated server url"}],"tags":[{"name":"Книги","description":"API для работы с книгами"},{"name":"Авторы","description":"API для работы с авторами"}],"paths":{"/book":{"get":{"tags":["Книги"],"summary":"Получение списка книг","operationId":"getList","parameters":[{"name":"text","in":"query","required":false,"schema":{"type":"string"}},{"name":"isbn","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"authorId","in":"query","required":false,"schema":{"type":"integer","format":"int64"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BookDto"}}}}}}},"post":{"tags":["Книги"],"summary":"Добавление новой книги","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewBookDto"}}},"required":true},"responses":{"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BookDto"}}}}}}},"/author":{"get":{"tags":["Авторы"],"summary":"Получение списка авторов","operationId":"getList_1","parameters":[{"name":"text","in":"query","required":false,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AuthorDto"}}}}}}},"post":{"tags":["Авторы"],"summary":"Добавление нового автора","operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewAuthorDto"}}},"required":true},"responses":{"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthorDto"}}}}}}},"/book/{bookId}":{"get":{"tags":["Книги"],"summary":"Получение информации о книги по идентификатору","operationId":"getById","parameters":[{"name":"bookId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BookDto"}}}}}},"delete":{"tags":["Книги"],"summary":"Удаление книги","operationId":"deleteById","parameters":[{"name":"bookId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"204":{"description":"No Content"}}},"patch":{"tags":["Книги"],"summary":"Изменение книги","operationId":"patch","parameters":[{"name":"bookId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewBookDto"}}},"required":true},"responses":{"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BookDto"}}}}}}},"/author/{authorId}":{"get":{"tags":["Авторы"],"summary":"Получение информации об авторе по идентификатору","operationId":"getById_1","parameters":[{"name":"authorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthorDto"}}}}}},"delete":{"tags":["Авторы"],"summary":"Удаление автора","operationId":"deleteById_1","parameters":[{"name":"authorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"204":{"description":"No Content"}}},"patch":{"tags":["Авторы"],"summary":"Изменение автора","operationId":"patch_1","parameters":[{"name":"authorId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewAuthorDto"}}},"required":true},"responses":{"409":{"description":"Conflict","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"404":{"description":"Not Found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiError"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthorDto"}}}}}}}},"components":{"schemas":{"ApiError":{"type":"object","properties":{"message":{"type":"string"},"reason":{"type":"string"},"status":{"type":"string"},"timestamp":{"type":"string","format":"date-time"}}},"NewBookDto":{"required":["authorId","isbn","name"],"type":"object","properties":{"name":{"maxLength":255,"minLength":3,"type":"string"},"isbn":{"type":"integer","format":"int64"},"authorId":{"type":"integer","format":"int64"}}},"AuthorDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"BookDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"isbn":{"type":"integer","format":"int64"},"author":{"$ref":"#/components/schemas/AuthorDto"}}},"NewAuthorDto":{"required":["name"],"type":"object","properties":{"name":{"maxLength":255,"minLength":3,"type":"string"}}}}}}